==> 01-data_extraction_nusc.ipynb <==
# Data Extraction

## Newcastle Special Collection & Archives Catalogs

***

**Table of Contents**

I. [For Entire Catalog](#for-entire-catalog)

II. [For Select Archival Collections](#for-select-archival-collections)

III. [Extract Descriptive Metadata](#extract-descriptive-metadata)

  - [For EAD XML](#for-ead-xml)
  - [For Non-EAD XML](#for-non-ead-xml)

IV. [Transformation](#transformation)

  - [Transcriptions](#transcriptions)
  - [Metadata](#metadata)
    - [EAD](#metadata-in-ead)
    - [Non-EAD](#metadata-not-in-ead)

***

# Libraries to extract metadata descriptions from XML data
import xml.etree.ElementTree as ET
from xml.etree.ElementTree import ParseError
from lxml import etree
import urllib.request
import time

# Libraries for data analysis
import pandas as pd
import numpy as np
import re

# Custom variables
import config
import ext_utils

# For writing data and creating directories
from pathlib import Path

## For Entire Catalog

nusc_url = "https://specialcollections.ncl.ac.uk/;oai?verb=ListRecords&metadataPrefix=oai_dc&set=oai:virtual:top-level-records"

id_tag = "{http://purl.org/dc/elements/1.1/}identifier"  # appears in browser as: <dc:identifier>
df = ext_utils.getTextFromTag(True, nusc_url, id_tag)
print(df.shape)
df.head()

eadids = df.eadid.unique()
eadids.sort()
print(eadids)

urls = list(df["url"])
xml_path_regex = "https?:\/\/\S*downloads\/exports\/ead\S*\.xml"
ead_xml_paths = []
for url in urls:
    response = urllib.request.urlopen(url)
    content = response.read().decode("utf-8")
    response.close()
    ead_xml_paths += re.findall(xml_path_regex, content)
print(len(ead_xml_paths))

## For Select Archival Collections

Read the CSV file with XML path data for the Newcastle University Special Collections archival catalog.

df_xml = pd.read_csv(config.xml_paths)
row_count = df_xml.shape[0]
col_count = df_xml.shape[1]
print("rows:",row_count, "| columns:",col_count)
# df_xml.head(3)
df_xml

print(list(df_xml["xml_path"])[5])

ead_xml_paths = [df_xml.iloc[i,1] for i in range(0,row_count) if df_xml.iloc[i,2] == "EAD"]
other_xml_paths = [df_xml.iloc[i,1] for i in range(0,row_count) if df_xml.iloc[i,2] != "EAD"] 
assert(ead_xml_paths[0] == df_xml.iloc[0,1]), "The first EAD XML list value should be the first XML path in the DataFrame."
assert(other_xml_paths[0] == df_xml.iloc[1,1]), "There should only be one non-EAD XML path, that of the Bell Archive."

## Extract Descriptive Metadata



### For EAD XML

Then, at all levels of the archival hierarchy, from the fonds (i.e., `<archdesc>`) down to the item level (e.g., `<c level="...">` for `"subfonds"`, `"subseries"`, `"file"`, `"item"`, etc.), extract the text between XML tags with the specified names, which are the descriptions in the metadata fields with those names.  

By default, the function we'll use, `extractMetadata()`, will look for text between the following tags:
* EADID: `<eadid>`
* Title: `<unittitle>`
* Identifier: `<unitid>`
* Date (of material?): `<unitdate>`
* Language of material: `<langmaterial>`
* Biographical / Historical: `<bioghist>`
* Scope and Contents: `<scopecontent>`
* Processing Information: `<processinfo>`
Optionally, you can specifiy your own list of tag (a.k.a. metadata field) names.

Generally, the Encoded Archival Description Identifier, or EADID (e.g., `<eadid>BXB<\eadid>`), will also be the first Unit ID of a fonds (e.g., `<unitid>BXB</unitid>`).

*Note 1: "Fonds" is the archival term for collection.  It is the top-most level of a single archive's (or archival collection's) hierarchy.*

*Note 2: The code for the `extractMetadata()` function can be found in the ext_utils.py file.*

To extract the metadata descriptions in the default list of tags for from all our EAD XML data, run the following code cell:

all_fonds = []
for xml_path in ead_xml_paths:
    print("Extracting archival metadata descriptions from", xml_path)
    fonds_list = ext_utils.extractMetadata(xml_path)
    all_fonds += [fonds_list]
print("Done!")

assert len(all_fonds) == len(ead_xml_paths), "The total number of fonds from which metadata descriptions were extracted should equal the number of EAD XML paths"

### For non-EAD XML
Extract metadata descriptions from the custom (non-EAD) XML data:

# Define the tags to extract text from
# Note: every item has a unique 'node_id'

metadata_tags = [  # for metadata descriptions
    "node_id", "title", "description", "creator", "creation-date", 
    "recipient", "language", "extent-and-medium", "country-and-region"
    ]

# transcription_tags = ["node_id", "type", "iiif-manifest", "transcription"] # For digitized text


more_data = True
resumption_token_prefix = "/export?resumptionToken="
errored_tokens, error_messages = [], []
text_list = []
for xml_path in other_xml_paths:
    next_xml_path = xml_path
    while more_data:
        print("Current XML path:", next_xml_path)
        # Create an ElementTree tree and get the tree's root
        content = urllib.request.urlopen(next_xml_path)

        try:

            parser = etree.XMLParser(recover=True)#, encoding="utf-8")  # Use recover to handle malformed XML data
            xmlTree = etree.parse(content, parser)
            root = xmlTree.getroot()

            # Extract metadata descriptions from the below-specified fields in the tree
            tags = transcription_tags
            d_descs = dict.fromkeys(tags, "")
            
            for child in xmlTree.iter():
                tag = child.tag
                if tag in tags:
                    all_text = ext_utils.getAllText(child, tag)
                    d_descs[tag] = all_text
                # Upon reaching a new item tag, add the previous item's extracted descriptions
                # to the metadata list and reset the descriptions dictionary
                elif (tag == "item") and (d_descs["node_id"] != ""):
                    text_list += [d_descs]
                    d_descs = dict.fromkeys(tags, "")
                else:
                    continue

            # Add the final item's descriptions
            text_list += [d_descs]

            # If there's a resumption token, use it to continue extracting data, 
            # otherwise there is no more data to extract
            another_token = root.find("resumptionToken")
            if not another_token is None:
                resumption_token = another_token.text
            else:
                more_data = False

        except ParseError as error_message:

            # Record the error
            errored_token = int(re.search("[0-9]+", resumption_token)[0])
            errored_tokens += [errored_token]
            error_messages += [error_message]
            print(error_message)

            # Try the next node ID for a resumption token
            next_token = errored_token + 1
            resumption_token = resumption_token_prefix + str(next_token)

        next_xml_path = xml_path.replace("/export/", resumption_token)

print(len(text_list))

text_list[1]

if len(errored_tokens) > 0:
    error_df = pd.DataFrame.from_dict({"resumption_token": errored_tokens, "error_message": error_messages})
    print(error_df.shape)
else:
    print("No parse errors!")

## Transformation

extracted_dir = "data/extracted/"
Path(extracted_dir).mkdir(parents=True, exist_ok=True)

### Transcriptions

Export the extracted data into a CSV file and the extracted transcriptions only into Plain Text files, where each file's name is the node ID of the transcribed item.

transcription_df = pd.DataFrame.from_dict(text_list)
print(transcription_df.shape)
transcription_df.head()

transcription_df.to_csv(extracted_dir + "bell_transcription_data.csv")

bell_transcriptions_dir = "data/extracted/GB_transcriptions/"
Path(bell_transcriptions_dir).mkdir(parents=True, exist_ok=True)

for item in text_list:
    node_id = item["node_id"]
    file_name = "GB_{}.txt".format(node_id)
    with open(bell_transcriptions_dir + file_name, "w", encoding="utf-8") as f:
        f.write(item["transcription"])
    f.close()
print("Wrote", len(text_list), "files to", bell_transcriptions_dir)

Check what the contents of one of the files looks like...

file_path = bell_transcriptions_dir + "GB_69571.txt"
with open(file_path) as f:
    full = f.read()
    print(full[:100])  # print the first 100 characters of the file
    f.close()

Looks good!

### Metadata

Store the extracted data in a DataFrame.

#### Metadata in EAD

metadata_df = pd.DataFrame()
for fonds_list in all_fonds:
    df = pd.DataFrame.from_records(fonds_list)
    metadata_df = pd.concat([metadata_df, df])
print(metadata_df.shape)

metadata_df.head()

Forward-fill the values in the `eadid` column so all rows have an EADID.

metadata_df["eadid"] = metadata_df["eadid"].replace({"":np.nan})
metadata_df["eadid"] = metadata_df["eadid"].ffill()
metadata_df.head()

metadata_df.tail()

f = "nusc_ead_all_fonds.csv" #"newcastle_archival_metadata_sample.csv"
metadata_df.to_csv(extracted_dir + f)

#### Metadata not in EAD

nonead_metadata_df = pd.DataFrame()
nonead_metadata_df = pd.DataFrame.from_records(text_list)
# The non-EAD formatted data is all from the Gertrude Bell Archive, which has the EADID "GB"
nonead_metadata_df.insert(0, "eadid", ["GB"]*(nonead_metadata_df.shape[0]))
print(nonead_metadata_df.shape)
nonead_metadata_df.head()

node_ids = list(nonead_metadata_df.node_id)
assert "69571" in node_ids

# nonead_metadata_df.to_csv(extracted_dir + "bell_archival_metadata_nonead.csv")

nonead_metadata_df.to_csv(extracted_dir + "complete_bell_archival_metadata_nonead.csv")
