==> 02-analysis_gb_transcriptions.ipynb <==
# Analysis of Transcriptions of the Gertrude Bell Archive

#### Table of Contents
* [Summary Statistics](#summary-statistics)

import os
import pandas as pd
import numpy as np
import re

import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
nltk.download('punkt')
nltk.download('punkt_tab')
from nltk.corpus import PlaintextCorpusReader
nltk.download('wordnet')
from nltk.corpus import wordnet
nltk.download('stopwords')
from nltk.corpus import stopwords
from nltk.text import Text
from nltk.stem.porter import PorterStemmer
from nltk.probability import FreqDist
nltk.download('averaged_perceptron_tagger')
from nltk.tag import pos_tag
nltk.download('tagsets')  # part of speech tags
from nltk.draw.dispersion import dispersion_plot as displt

# dir = "data/extracted/GB_transcriptions/"  # TXT files of transcriptions - one per item
f = "data/extracted/bell_transcription_data.csv"

## Summary Statistics

df = pd.read_csv(f, index_col=0)
df.tail()

df.type.value_counts()

Exclude items for which there are no transcriptions (i.e. the value in the `transcription` column is `NaN`).

print("Before:", df.shape)
subdf = df.dropna(subset=['transcription'])
print("After:", subdf.shape)

subdf.type.value_counts()

Nearly all of the text-based material has been transcribed!  99.5% of diaries and 99.9% of letters have transcriptions.  Diaries and letters account for 32.0% of the Gertrude Bell Archive's items (the rest are photographs).

transcriptions = list(subdf.transcription)
sentences = [sent_tokenize(t) for t in transcriptions]
print(sentences[0])

tokens = [[word_tokenize(s) for s in item_sentences] for item_sentences in sentences]
print(tokens[0][0])

# total_sentences = sum(len(item_sentences) for item_sentences in sentences)
# print("Total sentences:", total_sentences)
sentence_counts = [len(item_sentences) for item_sentences in sentences]
subdf.insert(len(subdf.columns), 'sentence_count', sentence_counts)
print("Total sentences:", sum(count for count in sentence_counts))

# total_tokens = sum(sum(len(sentence_tokens) for sentence_tokens in item_tokens) for item_tokens in tokens)
token_counts = [sum(len(sentence_tokens) for sentence_tokens in item_tokens) for item_tokens in tokens]
subdf.insert(len(subdf.columns), 'token_count', token_counts)
print("Total tokens:", sum(token_counts))

print(np.mean(sentence_counts))
print(np.min(sentence_counts))
print(np.max(sentence_counts))
print(np.std(sentence_counts))

print(np.mean(token_counts))
print(np.min(token_counts))
print(np.max(token_counts))
print(np.std(token_counts))

subdf.head()

Save a new version of the CSV file with the sentence and token count columns.

subdf.to_csv(f)

subdf.describe()

subdf.loc[subdf.type == "letters"].describe()

subdf.loc[subdf.type == "diary"].describe()
