==> 02-preprocessing.ipynb <==
# Preprocessing

Preprocess the data for classification, combining the EAD-formatted and custom-formatted data into a single file.

# Custom filepaths and functions
import config, ext_utils

# Libraries for data, file, and model loading
import pandas as pd

Load the extracted archival metadata from the Newcastle University Special Collections catalog.

extracted_dir = "data/extracted/"
ead_f = "nusc_ead_all_fonds.csv"  #"newcastle_archival_metadata_sample.csv"
ead_df = pd.read_csv(extracted_dir + ead_f, index_col=0)
nonead_df = pd.read_csv(extracted_dir + "complete_bell_archival_metadata_nonead.csv", index_col=0)   #nonead_df = pd.read_csv(extracted_dir + "bell_archival_metadata_nonead.csv", index_col=0)
print(ead_df.shape, nonead_df.shape)

ead_df.eadid.unique()

Transform the data so all the text that will be classified is in a single column.

text_cols = ["unittitle", "bioghist", "scopecontent", "processinfo"]
ead_doc_df = ext_utils.consolidateText(ead_df, text_cols, "unitid")

text_cols = ["title", "description", "extent-and-medium"]
nonead_doc_df = ext_utils.consolidateText(nonead_df, text_cols, "node_id")

doc_df = pd.concat([ead_doc_df, nonead_doc_df])
doc_df.sort_values(["eadid", "rowid", "field", "doc"], inplace=True)
doc_df.head()

Create a unique identifier for each description:

doc_df = doc_df.reset_index()
doc_df = doc_df.drop(columns=["index"])
doc_df = doc_df.reset_index()
doc_df = doc_df.rename(columns={"index":"description_id"})
doc_df.head()

print(doc_df.shape)

Save the descriptions with the unique identifier column.

#f = "nusc_archival_descs_for_clf.csv" # for classification
f = "nusc_archival_descs_consolidated.csv" # EAD and non-EAD (i.e., Bell descriptions)
doc_df.to_csv(extracted_dir+f)
