==> 03-data_deduplication.ipynb <==
# Data Deduplication

Remove duplicate descriptions from the exported data for manual review.

# Libraries for data analysis
import pandas as pd
import re

# Custom variables
import config

# For reading and writing data
from pathlib import Path
import os

### Get the Files and Directory

unclf_dir = "data/extracted/for_review/unclassified/"
file_names = os.listdir(unclf_dir)
print(file_names)

# Use the dataset of extracted, standardized data to create a new, deduplicated file of Gertude Bell Archive data 
unclf_data_file = "data/extracted/nusc_archival_descs_for_clf.csv"

Create a directory for the deduplicated, unclassified descriptions.

dedup_dir = "data/extracted/for_review/unclassified/deduplicated"
Path(dedup_dir).mkdir(parents=True, exist_ok=True)

### Deduplicate the descriptions

def implodeDataFrame(df, cols_to_groupby):
    cols_to_agg = list(df.columns)
    for col in cols_to_groupby:
        cols_to_agg.remove(col)
    agg_dict = dict.fromkeys(cols_to_agg, lambda x: x.tolist())
    return df.groupby(cols_to_groupby).agg(agg_dict).reset_index().set_index(cols_to_groupby)

First, deduplicate the EAD-formatted descriptions.

row_count = 0
all_descs = []
for f in file_names:
    if "_unclassified.csv" in f:
        print("Reading", f)
        # Read the file as a DataFrame
        df = pd.read_csv(unclf_dir+f, index_col=0)

        # Remove the empty column (where manual reviewers will make notes)
        df = df.drop(columns=["gender_bias?"])
        imploded = implodeDataFrame(df, ["doc","eadid"]).reset_index()
        print(df.shape, imploded.shape)
        row_count += imploded.shape[0]
        all_descs += list(imploded["doc"])
        assert imploded.shape[0] < df.shape[0], "There should be fewer rows in the DataFrame after imploding it."

        # Add a new empty column for manual reviewers to record their notes
        empty_col = [""]*imploded.shape[0]
        imploded.insert(len(imploded.columns), "gender_bias?", empty_col)
        
        # Reorder the columns
        imploded = imploded[["doc", "gender_bias?", "eadid", "description_id", "rowid", "field"]]

        # Write the imploded DataFrame into the newly created directory
        imploded.to_csv(dedup_dir+"/"+f)
print("Total rows:", row_count)

print(os.listdir(dedup_dir))

word_count = 0
for desc in all_descs:
    word_count += len(desc)
print("Total words:", word_count)  # about 2 million

Next, deduplicate the custom formatted descriptions (for the Gertrude Bell Archive).

df = pd.read_csv(unclf_data_file, index_col=0)
df = df.loc[df.eadid == "GB"]
print(df.shape)
df.head()


# Implode the DataFrame of Bell descriptions, grouping the data by doc (description)
imploded = implodeDataFrame(df, ["doc","eadid"]).reset_index()
print(df.shape, imploded.shape)
row_count = imploded.shape[0]
all_descs = list(imploded["doc"])
assert imploded.shape[0] < df.shape[0], "There should be fewer rows in the DataFrame after imploding it."

# Add a new empty column for manual reviewers to record their notes
empty_col = [""]*imploded.shape[0]
imploded.insert(len(imploded.columns), "gender_bias?", empty_col)

# Reorder the columns
imploded = imploded[["doc", "gender_bias?", "eadid", "description_id", "rowid", "field"]]

print("Total rows:", row_count)

word_count = 0
for desc in all_descs:
    word_count += len(desc)
print("Total words:", word_count)  # about 60k, making the grand total about 2.7 million words (2188536 + 601452 = 2789988)

imploded.head()

# Write the imploded DataFrame into the newly created directory
imploded.to_csv(dedup_dir+"/"+"GB_unclassified.csv")
