==> 04-coding_preparation.ipynb <==
# Coding Preparation
## Manual Review, part I

Create subsets of the unclassified and classified NUSC archival catalog metadata descriptions, one per collection, to share with each collection's designated expert for manual coding and classification evaluation.

# Library for data analysis
import pandas as pd

# Custom variables
import config

# For writing data and creating directories
from pathlib import Path

unclf_df = pd.read_csv("data/extracted/nusc_archival_descs_for_clf.csv", index_col = 0)
clf_osc_df = pd.read_csv("data/classified/baseline_osc_predictions.csv", index_col = 0)
clf_lc_osc_df = pd.read_csv("data/classified/lc-osc_predictions.csv")

unclf_dir = "data/extracted/for_review/unclassified/"
Path(unclf_dir).mkdir(parents=True, exist_ok=True)

clf_dir = "data/extracted/for_review/classified/"
Path(clf_dir).mkdir(parents=True, exist_ok=True)

# unclf_df.head()
# clf_osc_df.head()
clf_lc_osc_df.head()

# col_name = "gender_bias?"
# df = unclf_df
# file_suffix = "_unclassified.csv"
# file_dir = unclf_dir

col_name = "agree?"
df = clf_osc_df
file_suffix = "_classified.csv"
file_dir = clf_dir

df = df.rename(columns={"prediction":"prediction1"})

print(df.shape)
print(clf_lc_osc_df.shape)

index_cols = ["description_id", "eadid", "rowid", "field", "doc"]
df = df.join(clf_lc_osc_df[["description_id", "eadid", "rowid", "field", "doc", "os_predictions"]].set_index(index_cols), on=index_cols)
df = df.rename(columns={"os_predictions":"prediction2"})
print(df.shape)
df.head()

empty_col = [""]*df.shape[0]
df.insert(len(df.columns), col_name, empty_col)
df.tail()

eadids = list(df.eadid.unique())
print(eadids)

eadid_col = "eadid"

for eadid in eadids:
    subdf = df.loc[df[eadid_col] == eadid]
    subdf.to_csv(file_dir+eadid+file_suffix)
