==> 05-combine_manually_coded_data.ipynb <==
# Combine Manually Coded Data
## Manual Review, part II

import pandas as pd       # For data analysis
import numpy as np        # For numerical operations
from pathlib import Path  # For writing data and creating directories
import os                 # For reading data files
import config             # For accessing path variables

### I. Combine Manually Coded Data Files

Read all the CSV data files with manual coding.

coded_path = config.coded_data_path1
csvs = os.listdir(coded_path)
to_remove = ["all_manually_coded_data.csv", ".DS_Store"]
for f in to_remove:
    if f in csvs:
        csvs.remove(f)
print(csvs)

df1 = pd.DataFrame()
df2 = pd.DataFrame()
for csv in csvs:
    if "sample" in csv:
        subdf1 = pd.read_csv(coded_path + csv)
        df1 = pd.concat([df1, subdf1])
    else:
        subdf2 = pd.read_csv(coded_path + csv)
        df2 = pd.concat([df2, subdf2])

df1.rename(columns={
    "Gender bias":"gender_bias", "Unnamed: 1":"index"
    }, inplace=True)
df1.set_index("index", inplace=True)
df1.head(5)

df1.gender_bias.value_counts()

df2.rename(columns={
    "Unnamed: 0":"index", "If yeStereotype, Type?":"type", "gender_bias?":"gender_bias"
    }, inplace=True)
df2.set_index("index", inplace=True)
df2.fillna({"gender_bias": "n"}, inplace=True)
df2.fillna({"type": "NA"}, inplace=True)
df2.head(5)

Convert the last three columns' values from strings to lists.

col_names = ["description_id", "rowid", "field"]
for col_name in col_names:
    new_col = list(df2[col_name])
    new_col = [(col[1:-1]).replace("'", "").split(", ") for col in new_col]
    col_i = list(df2.columns).index(col_name)
    df2.drop(columns=[col_name], inplace=True)
    df2.insert(col_i, col_name, new_col)
df2.head()

Explode the DataFrame so there's only one description_id per row.

df2 = df2.explode(["description_id","rowid","field"])
df2.head()

Standardize the manual coding, with one column for gender bias, one column marking the type of gender bias, and and one column recording any notes about the observed gender bias.

df1["type"] = df1["gender_bias"].str.split(" - ", expand=True)[0]
df1["note"] = df1["gender_bias"].str.rsplit(" - ", expand=True)[1]
df1 = df1[["doc", "gender_bias", "type", "note", "eadid", "description_id", "rowid", "field"]]
df1.fillna({"gender_bias": "n"}, inplace=True)
df1.fillna({"type": "NA"}, inplace=True)
df1.fillna({"note": "NA"}, inplace=True)
df1.head()

df1.type.value_counts()

df1.note.value_counts()

gender_bias_values = list(df1.loc[df1.gender_bias != "n"].gender_bias)
df1 = df1.replace(to_replace=gender_bias_values, value="y")
df1.tail()

df2.insert(column="note", value="NA", loc=3)
df2.head()

df = pd.concat([df1, df2])
df.head()

df.gender_bias.value_counts()

df.type.value_counts()

df.shape

Replace the `type` column with two new columns to each record `"y"` or `"n"` for *Omission* and *Stereotype*.

type_list = list(df.type)
omission_manual = ["y" if "Omission" in t else "n" for t in type_list]
stereotype_manual = ["y" if "Stereotype" in t else "n" for t in type_list]
# print(type_list[100:150])
# print(omission_manual[100:150])
# print(stereotype_manual[100:150])

df.insert(column="omission", value=omission_manual, loc=2)
df.insert(column="stereotype", value=stereotype_manual, loc=3)
df.tail()

Export the combined manually coded data.

combined_filename = "all_manually_coded_data.csv"
df.to_csv(config.coded_data_path1+combined_filename)
