==> 06-baseline_os_clf.ipynb <==
# Classification

## Baseline Omission and Stereotype Classifier

***

**Table of Contents**
1. [Preprocessing](#1-preprocessing)
2. [Classification](#2-classification)
3. [Evaluation](#3-manual-review)
4. [Analysis: Description Lengths](#4-analysis)

import config, clf_utils
import re
import random
import numpy as np
import pandas as pd
from pathlib import Path
import joblib
from joblib import load
import nltk
nltk.download('punkt')
from nltk.tokenize import word_tokenize
from sklearn.multiclass import OneVsRestClassifier
from sklearn.linear_model import SGDClassifier
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
from sklearn.metrics import multilabel_confusion_matrix
from sklearn.metrics import precision_recall_fscore_support

### 1. Preprocessing

Load the extracted archival metadata from the Newcastle University Special Collections catalog.

# extracted_dir = "data/extracted/"
# doc_df = pd.read_csv(extracted_dir+"nusc_archival_descs_for_clf.csv", index_col=0)

manually_coded = config.coded_data_path1+"all_manually_coded_data.csv"
doc_df = pd.read_csv(manually_coded, index_col=0)
doc_df = doc_df.rename(columns={"gender_bias":"gender_bias_manual", "omission":"omission_manual", "stereotype":"stereotype_manual"})

doc_df.head()

doc_df.tail()

doc_df.shape  # All extracted data: (57947, 5) #complete Bell Archive metadata added about 25k more descriptions

There are 57,947 metadata descriptions extracted from the Newcastle University Special Collections archival catalog to classify for *Omission* and *Stereotype*.

There are 12,337 metadata descriptions that were manually coded for *Omission* and *Stereotype*.

### 2. Classification
Load the baseline Omission and Stereotype Classifier (OSC).

cvectorizer = joblib.load("models/transform_docs/count_vectorizer.joblib")
tfidf = joblib.load("models/transform_docs/tfidf_transformer.joblib")
mlb = joblib.load("models/transform_labels/mlb_targets_os.joblib")
clf = joblib.load("models/baseline_osc/sgd-svm_F-tfidf_T-os.joblib")

Classify the descriptions.

vectorized = cvectorizer.transform(doc_df["doc"])
X = tfidf.transform(vectorized)
y = clf.predict(X)

predictions = mlb.inverse_transform(y)

doc_df.insert(len(doc_df.columns), "baseline_prediction", predictions)
doc_df.head()

doc_df.baseline_prediction.value_counts()

# print("Proportion of descriptions classified:", (1489+909+314)/(55235+1489+909+314))  # about 4.7%
print("Proportion of manually coded descriptions classified:", (164+3+2)/(12168+164+3+2))  # about 1.4%

Export the classified data.

# filename = "baseline_osc_predictions.csv"
# doc_df.to_csv(config.classified_data_path+filename)

filename = "manually_coded_baselineosc.csv"
doc_df.to_csv(config.coded_and_classified+filename)

***
Deduplicate the descriptions and create a new export where each row has a unique doc-eadid-prediction combination.

doc_df_dedup1 = doc_df.drop_duplicates(subset=["eadid", "doc", "prediction"])
doc_df_dedup2 = doc_df.drop_duplicates(subset=["doc", "prediction"])
doc_df_dedup3 = doc_df.drop_duplicates(subset=["doc"])
print(doc_df_dedup1.shape)
print(doc_df_dedup2.shape)  # as expected, same as next
print(doc_df_dedup3.shape)  # as expected, same as previous

dedup_filename = "baseline_osc_predictions_deduplicated.csv"
doc_df_dedup1.to_csv(config.classified_data_path+dedup_filename)

doc_df_dedup1.prediction.value_counts()

print("Proportion of unique descriptions classified:", (1362+773+281)/(24819+1362+773+281) # about 8.87%

### 3. Manual Review
Export a selection of descriptions for manual review, including a combination of unclassified and classified descriptions, randomly selected to include, per collection equal parts descriptions that were unclassified, that were classified as *Omission*, that were classified as *Omission* and *Stereotype*, and that were classified as *Stereotype*.

eadid_col = "eadid"
eadids = list(doc_df_dedup1[eadid_col].unique())
print(eadids)

sample_path = config.classified_data_path+"sample_baseline/"
Path(sample_path).mkdir(parents=True, exist_ok=True)

file_suffix = "_sample.csv"
pred_col = "prediction"
sample_size = 0.05 # 5%
for eadid in eadids:
    subdf = doc_df_dedup1.loc[doc_df_dedup1[eadid_col] == eadid]
    unclf_docs = subdf.loc[subdf[pred_col] == ()] # descriptions without an Omission or Stereotype classification
    o_docs = subdf.loc[subdf[pred_col] == ('Omission',)] # descriptions with an Omission but no Stereotype classification
    s_docs = subdf.loc[subdf[pred_col] == ('Stereotype',)] # descriptions with a Stereotype but no Omission classification
    os_docs = subdf.loc[subdf[pred_col] == ('Omission', 'Stereotype')] # descriptions with an Omission and a Stereotype classification
    dfs = [unclf_docs, o_docs, s_docs, os_docs]

    all_samples = pd.DataFrame()
    for df in dfs:
        desc_id_list = list(df.description_id)
        if len(desc_id_list) > 0:
            sample_ids = random.sample(desc_id_list, int((len(desc_id_list))*sample_size))
        else:
            sample_ids = []
        sample = df.loc[df.description_id.isin(sample_ids)]
    
        all_samples = pd.concat([all_samples, sample])
    
    all_samples.sort_values(by=["description_id"])
    # print(eadid, "-", all_samples.shape)
    all_samples.to_csv(sample_path+eadid+file_suffix)
    print("Wrote",eadid+file_suffix+"!")

sample_path = config.classified_data_path+"sample_no_preds/"
Path(sample_path).mkdir(parents=True, exist_ok=True)

doc_df_dedup1.head()

file_suffix = "_sample_no_preds.csv"
pred_col = "prediction"
sample_size = 0.05 # 5%
for eadid in eadids:
    subdf = doc_df_dedup1.loc[doc_df_dedup1[eadid_col] == eadid]
    unclf_docs = subdf.loc[subdf[pred_col] == ()] # descriptions without an Omission or Stereotype classification
    o_docs = subdf.loc[subdf[pred_col] == ('Omission',)] # descriptions with an Omission but no Stereotype classification
    s_docs = subdf.loc[subdf[pred_col] == ('Stereotype',)] # descriptions with a Stereotype but no Omission classification
    os_docs = subdf.loc[subdf[pred_col] == ('Omission', 'Stereotype')] # descriptions with an Omission and a Stereotype classification
    dfs = [unclf_docs, o_docs, s_docs, os_docs]

    all_samples = pd.DataFrame()
    for df in dfs:
        df = df.drop(columns=[pred_col])
        desc_id_list = list(df.description_id)
        if len(desc_id_list) > 0:
            sample_ids = random.sample(desc_id_list, int((len(desc_id_list))*sample_size))
        else:
            sample_ids = []
        sample = df.loc[df.description_id.isin(sample_ids)]
    
        all_samples = pd.concat([all_samples, sample])
    
    all_samples.sort_values(by=["description_id"])
    # print(eadid, "-", all_samples.shape)
    all_samples.to_csv(sample_path+eadid+file_suffix)
    print("Wrote",eadid+file_suffix+"!")

### 4. Analysis
#### Description Lengths

unclf_docs = doc_df_dedup1.loc[doc_df_dedup1.prediction == tuple()]
unclf_docs.head()

clf_docs = doc_df_dedup1.loc[doc_df_dedup1.prediction != tuple()]
clf_docs.head()

print(unclf_docs.shape)
print(clf_docs.shape)

unclf_docs_list = list(unclf_docs.doc)
clf_docs_list = list(clf_docs.doc)
print(clf_docs_list[0])

unclf_token_counts = [len(word_tokenize(doc)) for doc in unclf_docs_list]
clf_token_counts = [len(word_tokenize(doc)) for doc in clf_docs_list]
print(unclf_token_counts[0])
print(clf_token_counts[0])

print("Minimum, average, standard deviation, median, and maximum token counts in UNCLASSIFIED descriptions:")
print(np.min(unclf_token_counts))
print(np.mean(unclf_token_counts))
print(np.std(unclf_token_counts))
print(np.median(unclf_token_counts))
print(np.max(unclf_token_counts))

print("Minimum, average, standard deviation, median, and maximum token counts in CLASSIFIED descriptions:")
print(np.min(clf_token_counts))
print(np.mean(clf_token_counts))
print(np.std(clf_token_counts))
print(np.median(clf_token_counts))
print(np.max(clf_token_counts))

o_docs = clf_docs.loc[clf_docs.prediction == ('Omission',)]
print(o_docs.shape)
o_docs.head()

os_docs = clf_docs.loc[clf_docs.prediction == ('Omission', 'Stereotype')]
print(os_docs.shape)
os_docs.head()

s_docs = clf_docs.loc[clf_docs.prediction == ('Stereotype',)]
print(s_docs.shape)
s_docs.head()
