==> 07-ling-os_clf.ipynb <==
# Classification

## Linguistic Classifier (LC) + Omission & Stereotype Classifier (OSC)

***

**Table of Contents**

* [Load the Data](#1-load-the-data)
* [Linguistic Classification](#2-linguistic-classification)
* [Omission and Stereotype Classification](#3-omission--stereotype-classification)
  * [Preprocessing](#31-preprocessing)
  * [Feature Extraction](#32-feature-extraction)
  * [Classification](#33-classification)
* [Analysis: Classified vs. Unclassified Descriptions](#4-analysis)


# Custom filepaths and functions
import config, clf_utils

# Libraries for data, file, and model loading
import pandas as pd
import random
import joblib
from joblib import load
import numpy as np
import nltk
nltk.download('punkt')
from nltk.tokenize import word_tokenize
from pathlib import Path

# Libraries for classification
from gensim.models import FastText
from gensim.utils import tokenize
from gensim import utils
from gensim.test.utils import get_tmpfile
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.feature_extraction.text import CountVectorizer
from skmultilearn.problem_transform import ClassifierChain
from sklearn.ensemble import RandomForestClassifier
from sklearn.multiclass import OneVsRestClassifier
from sklearn.linear_model import SGDClassifier
import scipy
import sklearn.metrics
from sklearn.metrics import confusion_matrix, multilabel_confusion_matrix
from sklearn.metrics import precision_recall_fscore_support

### 1. Load the Data

# extracted_dir = "data/extracted/"
# doc_df = pd.read_csv(extracted_dir+"nusc_archival_descs_for_clf.csv", index_col=0)

filename = "manually_coded_baselineosc.csv"
doc_df = pd.read_csv(config.coded_and_classified+filename)

doc_df.head()

text_cols = ["doc"]
token_df = clf_utils.getTokenDF(doc_df, text_cols)
token_df.head()

token_df.tail()

token_df.shape

### 2. Linguistic Classification
Classify the data using the Linguistic Classifier to create *Generalization*, *Gendered Pronoun*, and *Gendered Role* features to input into the Omission and Stereotype Classifier.

Load the models.

ft_model = FastText.load("models/embeddings/fasttextcbow100_lowercased/fasttext_cbow_100d.model")
mlb_ling = joblib.load("models/transform_labels/mlb_targets_ling.joblib")  # formerly named mlb_linlabels.joblib
ling_clf = joblib.load("models/ling_osc/cc-rf_F-fasttextcbow100_T-linglabels.joblib")

print("Classification targets:", mlb_ling.classes_)

Transform the token data into an embedding matrix to input into the classifier.

X = clf_utils.getFeatures(token_df, embedding_model=ft_model)

Classify the tokens.

y = ling_clf.predict(X)

predictions = mlb_ling.inverse_transform(y)
new_preds = [pred[0] if len(pred) > 0 else "O" for pred in predictions]

pred_col = "prediction"
token_df.insert(len(token_df.columns), pred_col, new_preds)
token_df.head()

token_df[pred_col].value_counts()

# token_df.to_csv(config.classified_data_path+"baseline_lc_predictions.csv")

token_df.to_csv(config.coded_and_classified+"manually_coded_baselinelc.csv")

### 3. Omission & Stereotype Classification
Classify the data using the Omission and Stereotype Classifier to assign *Omission* and *Stereotype* codes to NUSC archival metadata descriptions.

### 3.1 Preprocessing

# ling_preds_file = config.classified_data_path+"baseline_lc_predictions.csv" # token_df
# token_df = pd.read_csv(ling_preds_file, index_col=0)
# token_df.head()

# Group the data by description so it can be combined with the description data
df_features = clf_utils.implodeDataFrameUnique(token_df[["description_id", "token_id", pred_col]], ["description_id"])

# Deduplicate lists in description_id, token_id, and prediction columns
cols = ["token_id", pred_col]
for col in cols:
    old_col = list(df_features[col])
    new_col = []
    if col == pred_col:
        for values in old_col:
            if (len(values) > 1) and ("O" in values):
                values.remove("O")
            values_list = list(values)
            values_list.sort()
            new_col += [values_list]
    else:
        new_col = old_col
    col_i = list(df_features.columns).index(col)
    df_features = df_features.drop(columns=[col])
    df_features.insert(col_i, col, new_col)

df_features.head()

df_features.tail()

df_features[pred_col].value_counts()

### 3.2 Feature Extraction

Join the feature data (i.e., classifier-predicted *Gendered Pronoun*, *Gendered Role*, and *Generalization* labels) to the document data (i.e., NUSC archival catalog metadata descriptions).

df = df_features.join(doc_df.set_index("description_id"), on="description_id", how="right")

# Rename label columns for clarity
feature_col = "linguistic_prediction"
df = df.rename(columns={pred_col:feature_col})

# Remove "O" from linguistic prediction lists if other labels are present in the list
ling_pred = list(df[feature_col])
new_ling_pred = []
for pred in ling_pred:
    if len(pred) > 1:
        if "O" in pred:
            pred.remove("O")
    new_ling_pred += [pred]

df = df.drop(columns=[feature_col])
df.insert(3, feature_col, new_ling_pred)
df.head()

### 3.3 Classification

clf_os = joblib.load("models/ling_osc/sgd-svm_F-tfidf-ling_T-os.joblib")
mlb_os = joblib.load("models/transform_labels/mlb_targets_os.joblib")
cvectorizer = joblib.load("models/transform_docs/count_vectorizer.joblib")
tfidf_transformer = joblib.load("models/transform_docs/tfidf_transformer.joblib")

print(mlb_ling.classes_)
print(mlb_os.classes_)

docs = df["doc"]
vectorized = cvectorizer.transform(docs)
tfidf_matrix = tfidf_transformer.transform(vectorized)

features = mlb_ling.transform(df[feature_col])
X = scipy.sparse.hstack([tfidf_matrix, features])

y = clf_os.predict(X)
predictions = mlb_os.inverse_transform(y)
pred_col_name = "lcosc_prediction"
df.insert(len(df.columns), pred_col_name, predictions)
df.head()

df[pred_col_name].value_counts()

# print("Proportion of descriptions classified:", (2096+918+285)/(54648+2096+918+285)) # CHECK THIS: about 5.7% percent (1% more than found with baseline)
print("Proportion of manually coded descriptions classified:", (291+3+3)/(12040+291+3+3)) # about 2.4%

Save the classified data.

# filename = "lc-osc_predictions.csv"
# df.to_csv(config.classified_data_path+filename)

filename = "manually_coded_baselineosc_lcosc.csv"
df.to_csv(config.coded_and_classified+filename)

***
Deduplicate the descriptions and create a new export where each row has a unique doc-eadid-prediction combination.

df_dedup1 = df.drop_duplicates(subset=["eadid", "doc", "os_prediction"])
df_dedup2 = df.drop_duplicates(subset=["doc", "os_prediction"])
df_dedup3 = df.drop_duplicates(subset=["doc"])
print(df_dedup1.shape)
print(df_dedup2.shape)  # as expected, same as next
print(df_dedup3.shape)  # as expected, same as previous

dedup_filename = "lc-osc_predictions_deduplicated.csv"
df_dedup1.to_csv(config.classified_data_path+dedup_filename)

df_dedup1.os_prediction.value_counts()

print("Proportion of unique descriptions classified:", (1869+779+257)/(24330+1869+779+257)) # about 10.7%

## 4. Analysis
Calculate quantities of classified and unclassified descriptions.

unclf_docs = df_dedup1.loc[df_dedup1.os_prediction == tuple()]
unclf_docs.head()

clf_docs = df_dedup1.loc[df_dedup1.os_prediction != tuple()]
clf_docs.head()

print(unclf_docs.shape)
print(clf_docs.shape)

unclf_docs_list = list(unclf_docs.doc)
clf_docs_list = list(clf_docs.doc)
print(clf_docs_list[0])

unclf_token_counts = [len(word_tokenize(doc)) for doc in unclf_docs_list]
clf_token_counts = [len(word_tokenize(doc)) for doc in clf_docs_list]
print(unclf_token_counts[0])
print(clf_token_counts[0])

print("Minimum, average, standard deviation, median, and maximum token counts in UNCLASSIFIED descriptions:")
print(np.min(unclf_token_counts))
print(np.mean(unclf_token_counts))
print(np.std(unclf_token_counts))
print(np.median(unclf_token_counts))
print(np.max(unclf_token_counts))

print("Minimum, average, standard deviation, median, and maximum token counts in CLASSIFIED descriptions:")
print(np.min(clf_token_counts))
print(np.mean(clf_token_counts))
print(np.std(clf_token_counts))
print(np.median(clf_token_counts))
print(np.max(clf_token_counts))

clf_docs = clf_docs.reset_index()
unclf_docs = unclf_docs.reset_index()

o_docs = clf_docs.loc[clf_docs.os_prediction == ('Omission',)]
print(o_docs.shape)
o_docs.head()

os_docs = clf_docs.loc[clf_docs.os_prediction == ('Omission', 'Stereotype')]
print(os_docs.shape)
os_docs.head()

s_docs = clf_docs.loc[clf_docs.os_prediction == ('Stereotype',)]
print(s_docs.shape)
s_docs.head()
