==> 08-analysis_manual_vs_models.ipynb <==
# Data Analysis: Manual Codes vs. Model Classifications

Compare and contrast the manual coding of descriptions for *Omission* and *Stereotype* to the predictions of the Baseline Omission and Stereotype Classifier and the Omission and Stereotype Classifier with linguistic features.

***

**Table of Contents**

I. [Prepare Data for Manual Review, part 2](#i)

II. [Analyze Agreement, part 1](#ii)

   - [Manual vs. Baseline OSC](#ii-i)
   - [Manual vs. LCOSC](#ii-ii)
   - [Baseline OSC vs. LCOSC](#ii-iii)

***

# For custom variables and functions
import clf_utils
import config

# For data analysis
import pandas as pd
import numpy as np

# For reading and writing files and directories
import os
from pathlib import Path
import joblib
from joblib import load

# For evaluation of classification/coding
import sklearn.metrics
from sklearn.metrics import confusion_matrix, multilabel_confusion_matrix, ConfusionMatrixDisplay
from sklearn.metrics import precision_recall_fscore_support

Read manually coded and classified data.

dir = config.coded_and_classified
f = "manually_coded_baselineosc_lcosc.csv"
df = pd.read_csv(dir+f)
df.head(2)

Clean the model predictions' values.

df["baseline_prediction"] = df["baseline_prediction"].apply(lambda x: x.strip("()"))
df["baseline_prediction"] = df["baseline_prediction"].apply(lambda x: x.replace("'", ""))
df["baseline_prediction"] = df["baseline_prediction"].apply(lambda x: x.strip(","))
df["baseline_prediction"] = df["baseline_prediction"].apply(lambda x: x.split(", "))
df.baseline_prediction.value_counts()

df["lcosc_prediction"] = df["lcosc_prediction"].apply(lambda x: x.strip("()"))
df["lcosc_prediction"] = df["lcosc_prediction"].apply(lambda x: x.replace("'", ""))
df["lcosc_prediction"] = df["lcosc_prediction"].apply(lambda x: x.strip(","))
df["lcosc_prediction"] = df["lcosc_prediction"].apply(lambda x: x.split(", "))
df.lcosc_prediction.value_counts()

df["type"] = df["type"].replace("Sterotype", "Stereotype")
df["type"] = df["type"].replace("Stereotype and Omission", "Omission, Stereotype")
df["type"] = df["type"].fillna("")
df["type"] = df["type"].apply(lambda x: x.split(", "))
df.type.value_counts()

<a id="i"></a>
### I. Prepare Data for Manual Review, part 2

subdf = df.drop(columns=["token_id", "index", "linguistic_prediction", "gender_bias_manual", "omission_manual", "stereotype_manual"])
subdf = subdf.rename(columns={"type":"first_manual_code", "note":"first_note"})
subdf = subdf[[
    "doc", "baseline_prediction", "lcosc_prediction", "first_manual_code", "first_note",
      "eadid", "rowid", "field", "description_id"
    ]]
subdf.insert(1, "second_manual_code", np.nan)
subdf.insert(2, "second_note", np.nan)
subdf.tail(2)

subdf.head()

subdf.to_csv(dir+"manually_coded_and_classified.csv")

new_dir1 = "data/manually_coded_part2/group1/"
new_dir2 = "data/manually_coded_part2/group2/"
Path(new_dir1).mkdir(parents=True, exist_ok=True)
Path(new_dir2).mkdir(parents=True, exist_ok=True)

# eadids1 = ["BP", "HL", "THS"]
# eadids2 = ["BXB", "CHE", "OBR", "SH", "SW", "WCT"]
# subdf1 = subdf[subdf["eadid"].isin(eadids1)]
# subdf2 = subdf[subdf["eadid"].isin(eadids2)]
subdf3 = subdf[subdf["eadid"] == "CPT"]

# subdf1.to_csv(new_dir1+"manually_coded_and_classified.csv")
# subdf2.to_csv(new_dir2+"manually_coded_and_classified.csv")
subdf3.to_csv(new_dir2+"manually_coded_and_classified_CPT.csv")

<a id="ii"></a>
### II. Analyze Agreement, part 1
##### Compare manual codes to model classifications

Standardize manually coded and classified columns and values, creating columns for `omission_baseline`, `stereotype_baseline`, `omission_lcosc`, and `stereotype_lcosc`, and noting the presence of a code with `'y'` and the absence of a code with `'n'`.

cols = ["baseline_prediction", "lcosc_prediction"]
for col in cols:
    df_col = list(df[col])
    omission_clfs = ["y" if "Omission" in pred else "n" for pred in df_col]
    stereotype_clfs = ["y" if "Stereotype" in pred else "n" for pred in df_col]
    if "baseline" in col:
        suffix = "_baseline"
    else:
        suffix = "_lcosc"
    df.insert(column="omission"+suffix, value=omission_clfs, loc=(len(df.columns)))
    df.insert(column="stereotype"+suffix, value=stereotype_clfs, loc=(len(df.columns)))
df.head(2)

mlb = joblib.load("models/transform_labels/mlb_targets_os.joblib")

y_manual = mlb.transform(df["type"])
y_baseline = mlb.transform(df["baseline_prediction"])
y_lcosc = mlb.transform(df["lcosc_prediction"])
# print(y_baseline[190:200])

<a id="ii-i"></a>
##### i. Agreement: Manual vs. Baseline OSC

matrix = multilabel_confusion_matrix(y_manual, y_baseline, labels=[0,1])
print(matrix)

tn = matrix[:, 0, 0]  # True negatives
fn = matrix[:, 1, 0]  # False negatives
tp = matrix[:, 1, 1]  # True positives
fp = matrix[:, 0, 1]  # False positives
class_names = list(mlb.classes_)

[precision, recall, f_1, suport] = precision_recall_fscore_support(
    y_manual, y_baseline, beta=1.0, zero_division=0, labels=[0,1]
)

baseline_agmt_df = pd.DataFrame({
    "labels":class_names, "true_neg":tn, "false_neg":fn, "true_pos":tp, "false_pos":fp,
    "precision":precision, "recall":recall, "f_1":f_1
})
baseline_agmt_df

print("Macro scores:")
print(baseline_agmt_df[["precision", "recall", "f_1"]].mean())

precision, recall, F1 = clf_utils.precisionRecallF1(sum(tp), sum(fp), sum(fn))
print("Micro scores:")
print("precision \t", precision)
print("recall \t\t", recall)
print("f_1 \t\t", F1)

<a id="ii-ii"></a>
##### ii. Manual vs. LCOSC

matrix = multilabel_confusion_matrix(y_manual, y_lcosc, labels=[0,1])
print(matrix)

tn = matrix[:, 0, 0]  # True negatives
fn = matrix[:, 1, 0]  # False negatives
tp = matrix[:, 1, 1]  # True positives
fp = matrix[:, 0, 1]  # False positives
class_names = list(mlb.classes_)

[precision, recall, f_1, suport] = precision_recall_fscore_support(
    y_manual, y_lcosc, beta=1.0, zero_division=0, labels=[0,1]
)

lcosc_agmt_df = pd.DataFrame({
    "labels":class_names, "true_neg":tn, "false_neg":fn, "true_pos":tp, "false_pos":fp,
    "precision":precision, "recall":recall, "f_1":f_1
})
lcosc_agmt_df

print("Macro scores:")
print(lcosc_agmt_df[["precision", "recall", "f_1"]].mean())

precision, recall, F1 = clf_utils.precisionRecallF1(sum(tp), sum(fp), sum(fn))
print("Micro scores:")
print("precision \t", precision)
print("recall \t\t", recall)
print("f_1 \t\t", F1)

<a id="ii-iii"></a>
##### iii. Agreement: Baseline OSC vs. LCOSC

matrix = multilabel_confusion_matrix(y_baseline, y_lcosc, labels=[0,1])
print(matrix)

tn = matrix[:, 0, 0]  # True negatives
fn = matrix[:, 1, 0]  # False negatives
tp = matrix[:, 1, 1]  # True positives
fp = matrix[:, 0, 1]  # False positives
class_names = list(mlb.classes_)

[precision, recall, f_1, suport] = precision_recall_fscore_support(
    y_baseline, y_lcosc, beta=1.0, zero_division=0, labels=[0,1]
)

osc_agmt_df = pd.DataFrame({
    "labels":class_names, "true_neg":tn, "false_neg":fn, "true_pos":tp, "false_pos":fp,
    "precision":precision, "recall":recall, "f_1":f_1
})
osc_agmt_df

print("Macro scores:")
print(osc_agmt_df[["precision", "recall", "f_1"]].mean())

precision, recall, F1 = clf_utils.precisionRecallF1(sum(tp), sum(fp), sum(fn))
print("Micro scores:")
print("precision \t", precision)
print("recall \t\t", recall)
print("f_1 \t\t", F1)
